requirements.txt
flask
line-bot-sdk
yfinance
google-genai
gunicorn